Sm/default

+++ src/main/java/swu/musling/diary/service/DiaryServiceImpl.java
@@ -4,16 +4,16 @@ import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
-import swu.musling.diary.dto.CreateDiaryRequestDto;
-import swu.musling.diary.dto.CreateDiaryResponseDto;
-import swu.musling.diary.dto.DiaryResponseDto;
-import swu.musling.diary.dto.EmotionResponseDto;
+import swu.musling.diary.dto.*;
 import swu.musling.diary.jpa.Diary;
 import swu.musling.diary.jpa.DiaryRepository;
 import swu.musling.diary.jpa.Recommendation;
@@ -26,6 +26,7 @@ import swu.musling.music.jpa.Music;
 import swu.musling.music.jpa.MusicRepository;
 
 import javax.persistence.EntityNotFoundException;
+import java.time.LocalDate;
 import java.time.Year;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -112,17 +113,74 @@ public class DiaryServiceImpl implements DiaryService {
 
     @Override
     @Transactional
-    public void deleteDiary(Long diaryId) { //일기 삭제
+    public void deleteDiary(Member member, Long diaryId) { //일기 삭제
         Diary diary = diaryRepository.findById(diaryId)
                 .orElseThrow(() -> new EntityNotFoundException("Diary with id " + diaryId + " not found"));
+
+        // 사용자 확인
+        if (!diary.getMember().getMemberId().equals(member.getMemberId())) {
+            throw new AccessDeniedException("You are not authorized to delete this diary");
+        }
+
         diaryRepository.delete(diary);
     }
 
     @Override
-    public DiaryResponseDto getDiary(Long diaryId) {
+    @Transactional(readOnly = true)
+    public DiaryResponseDto getDiary(Member member, Long diaryId) {
         Diary diary = diaryRepository.findById(diaryId)
                 .orElseThrow(() -> new EntityNotFoundException("Diary not found with id: " + diaryId));
+
+        // 사용자 확인
+        if (!diary.getMember().getMemberId().equals(member.getMemberId())) {
+            throw new AccessDeniedException("You are not authorized to delete this diary");
+        }
+
         return DiaryResponseDto.fromEntity(diary);
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public List<DiaryResponseDto> getDiariesByDate(Member member, LocalDate date) {
+        List<Diary> diaries = diaryRepository.findAllByMemberAndDate(member, date);
+        return diaries.stream()
+                .map(DiaryResponseDto::fromEntity)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public Page<DiaryResponseDto> getAllDiaries(Member member, Pageable pageable) {
+        Page<Diary> diariesPage = diaryRepository.findAllByMember(member, pageable);
+        return diariesPage.map(DiaryResponseDto::fromEntity);
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public EmotionCountResponseDto getEmotionCounts(Member member) {
+        //회원의 모든 일기 조회
+        List<Diary> diaries = diaryRepository.findAllByMember(member);
+        //감정별 일기 수집 및 개수 계산
+        Map<String, Long> emotionCounts = diaries.stream()  //일기를 스트림으로 변환
+                /*
+                collect(Collectors.groupingBy(Diary::getMood, Collectors.counting()))는 스트림의 각 Diary 객체에서 getMood() 메서드를 호출하여 감정(mood) 값을 가져옵니다.
+                groupingBy는 이 감정 값을 기준으로 일기를 그룹화하고, counting() 컬렉터는 각 그룹에 속하는 일기의 수를 세어 Map<String, Long> 형태로 결과를 반환합니다.
+                여기서 키(String)는 감정의 이름이고 값(Long)은 해당 감정을 가진 일기의 개수입니다.
+                 */
+                .collect(Collectors.groupingBy(Diary::getMood, Collectors.counting()));
+        /*
+        emotionCounts 맵의 엔트리셋(각각의 키-값 쌍) 중 값(Long)이 가장 큰 엔트리를 찾습니다. 이 값은 가장 자주 등장하는 감정의 빈도수를 나타냅니다.
+        getKey() 메서드는 이 엔트리의 키, 즉 가장 자주 등장하는 감정의 이름을 반환합니다.
+        emotionCounts.get(mostFrequentEmotion)는 이 감정의 이름을 사용하여 그 감정이 몇 번 등장했는지, 즉 빈도수를 찾습니다.
+         */
+        String mostFrequentEmotion = Collections.max(emotionCounts.entrySet(), Map.Entry.comparingByValue()).getKey();
+        Long count = emotionCounts.get(mostFrequentEmotion);
+
+        return EmotionCountResponseDto.builder()
+                .emotionCounts(emotionCounts)
+                .mostFrequentEmotion(mostFrequentEmotion)
+                .mostFrequentEmotionCount(count)
+                .build();
     }
 
     public List<String> getPreferredGenres(Member member) { //사용자 선호 장르
